requirements.txt
torch>=2.1.0
transformers>=4.41.0
datasets>=2.20.0
peft>=0.11.0
accelerate>=0.31.0
bitsandbytes>=0.43.0
unsloth>=2024.5.9
pyyaml>=6.0.1
scipy>=1.12.0
evaluate>=0.4.0
tqdm>=4.66.0
huggingface_hub>=0.23.0
